requirements.txt
Flask~=2.0.3
flask_socketio
playsound~=1.3.0
gtts~=2.2.4
websockets~=9.1
pyjwt~=1.7.1
infra-kess~=1.1.5

SQLAlchemy~=1.4.54
protobuf~=3.19.6
grpcio~=1.48.2
Flask-SocketIO~=5.4.1
Flask-Cors~=5.0.0
matplotlib~=3.3.4
requests~=2.27.1
objtyping~=0.5.4